chore: enable -Wall -Wextra for project code and add CI
Scope -Wall/-Wextra to project code only via an `orange_project_warnings`
INTERFACE library linked into every target; third-party/SDK include dirs are
attached as SYSTEM includes and vendored source files opt out with a
source-scoped -w. Vendored NvEncoder headers (reached through the current-file
directory, so -isystem cannot cover them) get `#pragma GCC system_header`.

Burn down all resulting project-code warnings (mechanical, behavior-preserving
fixes: unused params/vars, sign-compare casts, -Wreorder init-list ordering,
enum/int conditional, multi-line comment). -Wmissing-field-initializers is
suppressed as one class (~278 benign NVENC `{ NV_ENC_*_VER }` zero-inits).
-Werror is intentionally not enabled.

Fix the Debug-only ODR link error by making
EncoderPreprocessWorker::DEFAULT_ENCODER_ENTRY_POOL_SIZE static constexpr.

Add a minimal self-hosted CI workflow (release build+ctest, debug_tsan
stress/hammer tests) with a cancel-in-progress concurrency group.

+++ src/pose_worker.cpp
@@ -970,7 +970,7 @@ void PoseWorker::publish_pose_result_v2(
         object.confidence = pose ? static_cast<float>(pose->confidence) : frame.detection_confidence;
         object.label_id = 0;
         object.track_id = -1;
-        object.flags = publish_detection_bbox ? shaman_v2::kObjectHasBbox : 0;
+        object.flags = publish_detection_bbox ? static_cast<uint32_t>(shaman_v2::kObjectHasBbox) : 0u;
         if (pose) {
             object.flags |= shaman_v2::kObjectHasPose;
             object.keypoint_count = static_cast<uint32_t>(
